# protobuf: cache temp buffers