Alternate implementation of https://github.com/grpc/grpc-java/pull/2288