@buchgr 2016-09-21
I have considered that too, however 1) This is executed by an application thread. There might be lots of application threads and large messages using up lots of memory that can't be freed/reused until the thread dies. 2) Android?