@carl-mastrangelo 2016-09-21
The memory is contained entirely in this file, and never escapes.  There should be no memory pressure added.  If there are lots of threads, the threads are going to dominate, each with multi megabyte stacks.  

For android the worst that will happen is that the buffer is allocated on the heap, and retained by the thread local.  I think this would be strictly an improvement over the current behavior on Android.