@buchgr 2016-09-21
We keep up to 4MiB per thread. With hundreds of application threads, there is potential to waste significant memory...

I thought Android was memory constraint? I think it could look like a memory leak. Send a few large messages and the application will suddenly need much more memory that's never reclaimed.