@carl-mastrangelo 2016-09-21
I'm submitting this, to see the effect on the open source benchmarks, I can revert it if the memory implications are bad enough.  I expect the number of threads to be bounded in the majority of cases.